Gradleisation of Project
Originally set as a maven project, have converted this to a gradle
project.  Gradle enables faster build times over maven and can be
expanded with a greater set of tools.

In order to get this to build:

1.  Get the latest gradle at:  https://gradle.org/install/ and install
    manually
2.  Run gradle build ->  boom you're there!

To get working with existing Intellij project:

1.  Close project and delete it from Intellij start menu
2.  Import new Project
3.  Select the build.gradle file.  This will ensure Intellij creates the
    project as a gradle project.
4.  Run gradle build from side menu ->  boom you're there

Some troubles you could have with Intellij, include the message "Build failed with "Failed to notify progress listener."
If this happens:

1.  Try to change your Gradle settings in Intellij (Build,Execution,
    Deployment) -> Build Tools -> Gradle -> "Use Gradle From:" and
    change this to a specified location.  Then point to where your
    Gradle install was located e.g. C:/Gradle/gradle-7.0

2.  Upgrade Intellij.  Both Java 16 and Gradle 7 have some mismatches in
    older versions of Intellij, and as such build errors occur.  As of
    writing this should work with Intellij 2021.1, Gradle 7, and Java
    16

Leaving pom.xml in src code until entire team have migrated and built
successfully.